+++ localstack/services/cloudformation/engine/template_deployer.py
@@ -17,13 +17,13 @@ from localstack.services.cloudformation.deployment_utils import (
     is_none_or_empty_value,
     remove_none_values,
 )
+from localstack.services.cloudformation.engine import template_preparer
 from localstack.services.cloudformation.service_models import (
     KEY_RESOURCE_STATE,
     DependencyNotYetSatisfied,
     GenericBaseModel,
 )
 from localstack.utils.aws import aws_stack
-from localstack.utils.cloudformation import template_preparer
 from localstack.utils.collections import merge_recursive
 from localstack.utils.functions import prevent_stack_overflow, run_safe
 from localstack.utils.json import clone_safe, json_safe
