+++ localstack/services/cloudformation/engine/template_preparer.py
