+++ tests/unit/test_cloudformation.py
@@ -5,9 +5,9 @@ from localstack.services.cloudformation.deployment_utils import (
     PLACEHOLDER_AWS_NO_VALUE,
     remove_none_values,
 )
+from localstack.services.cloudformation.engine import template_deployer, template_preparer
 from localstack.services.cloudformation.models.stepfunctions import _apply_substitutions
 from localstack.services.cloudformation.provider import Stack
-from localstack.utils.cloudformation import template_deployer, template_preparer
 
 
 def test_resolve_references():
